Deploy/file_worker/requirements.txt
hiredis
celery[redis]
PyPDF2
reportlab
ftfy
pdfminer.six